requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation
pydantic==2.5.0

# Image processing
Pillow==10.1.0

# EasyOCR for Tamil text recognition
easyocr==1.7.0
opencv-python-headless==4.8.1.78

# HTTP requests
requests==2.32.4

# Environment variables
python-dotenv==1.0.0

numpy==1.26.4